Refactor BindingInformation and FunctionInformation to Common
