+++ azure-functions-powershell-library/BundledBindings/IBinding.cs
@@ -3,6 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
+using Common;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Management.Automation.Language;
