+++ azure-functions-powershell-library/BundledBindings/IInputBinding.cs
@@ -3,6 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
+using Common;
 using System.Management.Automation.Language;
 
 namespace AzureFunctions.PowerShell.SDK.BundledBindings
