+++ azure-functions-powershell-library/BundledBindings/IOutputBinding.cs
@@ -3,6 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
+using Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
