+++ azure-functions-powershell-library/Common/FunctionInformation.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-namespace AzureFunctions.PowerShell.SDK
+namespace Common
 {
     public class FunctionInformation
     {
