+++ azure-functions-powershell-library/GetFunctionsMetadataCommand.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-using AzureFunctions.PowerShell.SDK;
+using Common;
 using Microsoft.Azure.Functions.PowerShellWorker;
 using System.Management.Automation;
